+++ Oqtane.Application/Server/wwwroot/Modules/Templates/Internal/Client/Services/[Module]Service.cs
@@ -7,18 +7,6 @@ using Oqtane.Shared;
 
 namespace [Owner].Module.[Module].Services
 {
-    public interface I[Module]Service 
-    {
-        Task<List<Models.[Module]>> Get[Module]sAsync(int ModuleId);
-
-        Task<Models.[Module]> Get[Module]Async(int [Module]Id, int ModuleId);
-
-        Task<Models.[Module]> Add[Module]Async(Models.[Module] [Module]);
-
-        Task<Models.[Module]> Update[Module]Async(Models.[Module] [Module]);
-
-        Task Delete[Module]Async(int [Module]Id, int ModuleId);
-    }
 
     public class [Module]Service : ServiceBase, I[Module]Service
     {
